Change lexer_323() to accept an istream
Previously, lexer_323() accepted an ifstream. This makes testing much
harder, since we now have to make a temporary file just to run the
function, even though the function doesn't really care if the stream is
actually in a file.

This commit changes lexer_323() to accept an istream instead, which lets
us create a stringstream and pass that into lexer_323() trivially
without needing to touch the filesystem at all.

There is no good reason for why lexer_323() would want an ifstream in
the first place.