Fix test() crashing with out-of-bounds exception
Previously, we assume that location will always be within the length of
the tokens vector.

However, this is not the case. The procedure_* function might report a
missing token at the end of the input, meaning that location would be
equal to the length of the tokens.

This commit checks for that edge case.